requirements.txt
Pillow
PyQt6
pillow-heif
